move all public files to public folder
