+++ lib/csv_row_model/public/model.rb
