+++ lib/csv_row_model/public/model/file_model.rb
